Split answer and advice routes
